+++ src/docs/blog/self_documenting_apis_with_openapi/basic_route_test.kt
@@ -1,7 +1,6 @@
 package blog.self_documenting_apis_with_openapi
 
 import com.natpryce.hamkrest.assertion.assertThat
-import org.http4k.contract.contract
 import org.http4k.core.Method.GET
 import org.http4k.core.Request
 import org.http4k.hamkrest.hasBody
@@ -10,7 +9,6 @@ import org.junit.jupiter.api.Test
 class GreetingRouteTest {
     @Test
     fun `greets an adult`() {
-        val app = contract { routes += basicRoute }
-        assertThat(app(Request(GET, "/greet/Bob/21")), hasBody("Hello Bob, would you like some beer?"))
+        assertThat(basicRoute(Request(GET, "/greet/Bob/21")), hasBody("Hello Bob, would you like some beer?"))
     }
 }