+++ CHANGELOG.md
@@ -5,6 +5,7 @@ changes with their rationale when appropriate:
 
 ### v3.165.0 (uncut)
 - [http4k-contract] Fix invalid OpenAPI2 when root and base path match. H/T @rgladwell
+- [http4k-contract] `ContractRoute` is now an `HttpHandler`, so no need to wrap contract routes in a `contract {}` to test them. H/T @rgladwell for the inspiration.
 
 ### v3.164.0
 - [all] Update various dependencies, including Kotlin to 1.3.41.